requirements.txt
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
configparser==5.1.0
psycopg2==2.9.1
PyJWT==2.3.0
bcrypt==3.2.0
Werkzeug~=2.0.2
pandas==1.3.4
Flask-Cors==3.0.10
flask-jwt-extended==4.3.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
flask-accepts==0.18.4
